add unit tests file to github split the app on multiple files
