+++ pyproject.toml
@@ -12,6 +12,7 @@ SQLAlchemy = "^1.4.32"
 pycoingecko = "^2.2.0"
 app = "^0.0.1"
 APScheduler = "^3.9.1"
+functions = "^0.7.0"
 
 [tool.poetry.dev-dependencies]
